requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.5.0
python-dateutil==2.8.2
loguru==0.7.2
psycopg2-binary==2.9.9
dateparser==1.1.8
